fix:kbar reopen state
